set default last_read_at

+++ app/models/fav.rb
@@ -13,7 +13,7 @@ class Fav < ApplicationRecord
   def destroy_notification_if_zero_fav
     if score.favs_count == 1
       notification = Notification.find_by(title: score.token)
-      notification.destroy
+      notification.destroy if notification.present?
     end
   end
 end
